Add tests, schemas and other bits
